refs #98 Move job package under pkg directory
