+++ .github/workflows/test.yml
@@ -19,4 +19,4 @@ jobs:
       - name: Testing
         run: |
           go mod download
-          go test ./...
+          go test ./pkg/...
